requirements.txt
Flask-Cors==3.0.10
Flask==2.0.1
PyMySQL==1.0.2
coloredlogs==15.0.1
flasgger==0.9.5
pytest-cov==3.0.0
uWSGI==2.0.19.1
